server/requirements.txt
channels>=4.0.0
daphne>=4.0.0
Django>=4.2.6
djangorestframework==3.14.0
django-filter==23.4
Pillow>=10.1.0